Initial build system and CI (#6)
Setup the build environment and initial CI. Closes
https://github.com/tenstorrent/tt-lang/issues/7

* Create initial directory structure that follows typical LLVM/MLIR
conventions
* Add CMake build support
* tt-mlir is built separately and included as a third-party package; its
settings and exported targets are available within tt-lang
* Add initial build and test workflows for Linux and Mac OSX (with
caching of tt-mlir). Since the Mac workflow is not using a container, it
also caches the LLVM build.
* Add a sample MLIR pipeline for testing the build, mostly a copy of the
equivalent one in tt-mlir

## TODO
* Set up `main` PR rules to require all workflows to pass before a PR
can be merged (requires repo admin)
* In a future PR, extend `ExternTTMLIR.cmake` to automatically clone,
configure, and build tt-mlir (and if necessary, the toolchain).