[fix] Eliminate unnecessary llvm rebuilds on macOS CI (#29)
The github cache action does not work correctly when a secondary key was
used to restore the cache, causing unnecessary extra rebuilds of llvm.
This PR fixes that problem by checking for the existence of the
(restored) toolchain instead of the cache success variable.

In addition, the
https://github.com/tenstorrent/tt-lang/actions/workflows/call-build-tt-mlir-fetchcontent.yml
workflow (which runs nightly) now supports manual workflow triggers with
a user-specified tt-mlir SHA.